# [DO NOT MERGE] storage: benchmark difference between purego and cgo iteration