--- a/pkg/server/server.go
+++ b/pkg/server/server.go
@@ -1110,6 +1110,11 @@ If problems persist, please see ` + base.DocsURL("cluster-setup-troubleshooting.
 	s.sqlExecutor.Start(ctx, &s.adminMemMetrics, distSQLPlanner)
 	s.distSQLServer.Start()
 
+	if len(s.engines) > 1 {
+		panic("use only one engine")
+	}
+	s.distSQLServer.ServerConfig.Engine = s.engines[0]
+
 	atomic.StoreInt32(&s.serveNonGossip, 1)
 
 	log.Infof(ctx, "starting %s server at %s", s.cfg.HTTPRequestScheme(), unresolvedHTTPAddr)