NOTE: All numbers collected on a one-node cluster running on my laptop.

While investigating `COUNT(*)` performance (https://github.com/cockroachdb/cockroach/issues/19288), the throughput of various layers was measured (fb05b62) while performing `SELECT COUNT(*) FROM tpch.lineitem`. The throughput found for the `rocksdb` layer (which is simple iteration over the keys/values in `tpch.lineitem`) reported `~95MB/s`. This was a long way off from the `~500MB/s` measured through [`BenchmarkMVCCComputeStats_RocksDB`](https://github.com/cockroachdb/cockroach/blob/03a829c5f069a75a351edef36fe255055e73a6b1/pkg/storage/engine/bench_rocksdb_test.go#L75). The difference between these two iterations is that for the first one, we cross the cgo boundary once for every key/value pair while we only cross the cgo boundary once while computing stats, as the iteration is done in C++.

This warranted further investigation, so we set about measuring the cgo overhead by ingesting a `tpch.lineitem` sst and reading it with a `leveldb` pure go iteration and our normal `rocksdb` iteration (included in this PR), and got a pretty noticeable difference (the size of a key/value pair is `~200` bytes):
```
$ benchstat cgo purego
name             old time/op    new time/op     delta
MVCCIteration-8     511ms ± 4%      139ms ± 2%   -72.84%  (p=0.000 n=19+18)

name             old speed      new speed       delta
MVCCIteration-8  47.8MB/s ± 4%  176.2MB/s ± 2%  +268.82%  (p=0.000 n=20+18)
```
Although the purego iteration doesn't hit the `500MB/s` number, the throughput really depends on the size of the keys so is not comparable (for example, here is a run with an sst file where each key/value was `~700 bytes`):
```
BenchmarkMVCCIteration/Cgo-4         30      38456904 ns/op    1093.95 MB/s
BenchmarkMVCCIteration/Purego-4      50      27323478 ns/op    1539.69 MB/s
```
By moving our iteration at the MVCCScan level into C++, we could reduce our cgo overhead for scans and increase the throughput by a significant amount.
@arjunravinarayan @petermattis @jordanlewis @bdarnell @nvanbenschoten