@petermattis 2017-10-27
> By moving our iteration at the MVCCScan level into C++, we could reduce our cgo overhead for scans and increase the throughput by a significant amount.

I agree, though there are significant challenges to doing this due to the cgo calling constraints and the restriction that only  C code can be called from Go, not C++. 

---

Review status: 0 of 8 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19602#-:-KxUaCoiJh16-nNHMgAk:b2l6csv)*
<!-- Sent from Reviewable.io -->