@petermattis 2017-10-27
> I agree, though there are significant challenges to doing this due to the cgo calling constraints and the restriction that only  C code can be called from Go, not C++.

To elaborate: the challenge is in passing the results of the scan from C++ back through Cgo to Go. On possibility is that we return the results in RocksDB WriteBatch format. This is independent of changing ScanResponse to use WriteBatch format. The benefit of using WriteBatch is that it is a single chunk of memory and not a complex data structure. 

---

Review status: 0 of 8 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19602#-:-KxUePfdU7o1VmUoUlj9:b9va3au)*
<!-- Sent from Reviewable.io -->