@bdarnell 2017-10-28
We could also reduce cgo boundary crossing while keeping the current structure by prefetching multiple k/v pairs whenever we call into rocksdb. It may be tricky to take advantage of these prefetched keys since we do a lot of `Seek()` calls and not just `Next()`.